Initialize dev branch e2e testing process

+++ examples/dbt/mrr-playbook/macros/drop_table.sql
@@ -1,6 +1,12 @@
-{%- macro drop_table(table_name) -%}
-    {%- set drop_query -%}
-        DROP TABLE IF EXISTS {{ target.schema }}.{{ table_name }} CASCADE
-    {%- endset -%}
+{%- macro drop_table(table_name, conn_type) -%}
+    {% if conn_type != 'bigquery' %}
+      {%- set drop_query -%}
+          DROP TABLE IF EXISTS {{ target.schema }}.{{ table_name }} CASCADE
+      {%- endset -%}
+    {% else %}
+      {%- set drop_query -%}
+          DROP TABLE IF EXISTS {{ target.schema }}.{{ table_name }}
+      {%- endset -%}
+    {% endif %}
     {% do run_query(drop_query) %}
 {%- endmacro -%}
